Small changes
changelog_client : has been added and you will be able to view the
changes of the client.

changelog_server : has been added and you will be able to view some
changes and future content we have to add to the server. This only
contains basic information. We will surprise you with some bonus
content, don't worry but we will not write it in the changelog as a
future update ;)
Server will be online as soon as the basic's are done :)

Stay tuned!